requirements.txt
dj-database-url==0.4.2
Django==1.11.4
gunicorn==19.7.1
whitenoise==3.3.0
psycopg2==2.6.2
facebook-sdk==2.0.0
